--- app/src/main/kotlin/jp/co/yumemi/android/codecheck/OnItemClickListener.kt
@@ -1,5 +1,0 @@
-package jp.co.yumemi.android.codecheck
-
-interface OnItemClickListener {
-    fun itemClick(item: RepositoryInfo)
-}
