+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/adapter/CustomAdapter.kt
@@ -1,9 +1,11 @@
-package jp.co.yumemi.android.codecheck
+package jp.co.yumemi.android.codecheck.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
+import jp.co.yumemi.android.codecheck.R
+import jp.co.yumemi.android.codecheck.RepositoryInfo
 
 class CustomAdapter(
     private val itemClickListener: OnItemClickListener,
