+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/adapter/OnItemClickListener.kt
@@ -1,0 +1,7 @@
+package jp.co.yumemi.android.codecheck.adapter
+
+import jp.co.yumemi.android.codecheck.RepositoryInfo
+
+interface OnItemClickListener {
+    fun itemClick(item: RepositoryInfo)
+}
