+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/adapter/ViewHolder.kt
@@ -1,8 +1,10 @@
-package jp.co.yumemi.android.codecheck
+package jp.co.yumemi.android.codecheck.adapter
 
 import android.view.View
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import jp.co.yumemi.android.codecheck.R
+import jp.co.yumemi.android.codecheck.RepositoryInfo
 
 class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
     private val nameText = view.findViewById<TextView>(R.id.repositoryNameView)
